Powershell tools
Basic tools for local powershell exploitation
This set is designed to be used with a usb pen drive or installer of
some description.

This is provided as is and is by no means complete or near to be
finished.

You will need to change to file output paths in the script for the time
being until I assign a variable for where they should go.